+++ main.py
@@ -2,7 +2,8 @@ from dev.application import *
 
 gm_microscope = MyMicroscope("Goeppert Mayer",
     gm_camera = MyCamera("Camera CS165MU", "28939"),
-    gm_spectrometer = MySpectrometer("Spectrometer", "QEP06226"),
+    gm_spectrometer = MySpectrometer("Spectrometer1", "QEP06226"),
+    gm_spectrometer2 = MySpectrometer("Spectrometer2", "QEP06226"),
     gm_white_light = MyWhiteLight("White light", "xxxx"),
     gm_laser_750 = MyLaser("Laser 750nm", "xxxx"),
     gm_laser_550 = MyLaser("Laser 550nm", "xxxx"),
@@ -17,5 +18,5 @@ lm_microscope = MyMicroscope("Lisa Meitner",
     lm_filter_12 = MyFilter("Filter 12", "xxxx"),
     lm_platform = MyPlatform("Platform", "xxxx"))
 
-app = MyApp("V0.2.1", (1200, 700), gm_microscope, lm_microscope)
+app = MyApp("V0.2.2", (1200, 700), gm_microscope, lm_microscope)
 app.mainloop()